rename fjson folder
